+++ attributes/default.rb
@@ -1,4 +1,4 @@
-default[:niet][:install_source] = "https://github.com/willbryant/niet/tarball"
+default[:niet][:install_source] = "https://github.com/willbryant/niet/archive"
 default[:niet][:tag]            = "master"
 default[:niet][:artifact_type]  = "tar.gz"
 default[:niet][:safe_install]   = true